+++ src/aws_ha_script/ngfw_utils.py
@@ -1,11 +1,25 @@
 import re
 import subprocess
-import sys
 from typing import Optional
 from aws_ha_script.aws_utils import EC2ResourceType, get_config_tag_value, get_metadata_value
 from aws_ha_script.config import HAScriptConfig
+from aws_ha_script.exceptions import HAScriptConfigError
 from aws_ha_script.log_utils import logger
-from aws_ha_script.smc_events import send_error_to_smc
+
+
+def is_instance_type(config, instance_id_type: str) -> bool:
+    instance_id = get_metadata_value("instance-id")
+    if not config.primary_instance_id:
+        raise HAScriptConfigError("Missing primary_instance_id")
+    if not config.secondary_instance_id:
+        raise HAScriptConfigError("Missing secondary_instance_id")
+    if instance_id not in [config.primary_instance_id, config.secondary_instance_id]:
+        raise HAScriptConfigError(f"Instance id not configured correctly: {instance_id}")
+    if instance_id_type == "primary" and instance_id == config.primary_instance_id:
+        return True
+    if instance_id_type == "secondary" and instance_id == config.secondary_instance_id:
+        return True
+    return False
 
 
 def is_primary(config) -> bool:
@@ -13,13 +27,10 @@ def is_primary(config) -> bool:
 
     check if this engine is primary by comparing its own instance-id
     with the primary_instance_id defined in config.
+
+    :param config: configuration from the main program
     """
-    instance_id = get_metadata_value("instance-id")
-    primary_instance_id = config.primary_instance_id
-    if not primary_instance_id:
-        logger.critical("Config error: missing primary_instance_id")
-        sys.exit(1)
-    return primary_instance_id == instance_id
+    return is_instance_type(config, "primary")
 
 
 def is_secondary(config) -> bool:
@@ -27,18 +38,25 @@ def is_secondary(config) -> bool:
 
     check if this engine is secondary by comparing its own instance-id
     with the primary_instance_id defined in config.
+
+    :param config: configuration from the main program
     """
-    return not is_primary(config)
+    return is_instance_type(config, "secondary")
 
 
-def set_local_status(new_status: str) -> bool:
+def set_local_status(config, new_status: str) -> bool:
     """Change the node status ("offline" or "online")
 
+    :param config: configuration from the main program
     :param new_status: "offline" or "online"
     :return: True if successful, False otherwise.
     :raises: None
     """
     assert new_status == "online" or new_status == "offline"
+
+    if config.dry_run:
+        logger.warning("DRY-RUN: Do not change node status to %s.", new_status)
+        return True
 
     try:
         exit_status = subprocess.call(
@@ -46,7 +64,7 @@ def set_local_status(new_status: str) -> bool:
         )
         is_success = exit_status == 0
     except OSError:
-        logger.exception("Failed to change node status to %s", new_status)
+        logger.exception("Failed to change node status to %s.", new_status, exc_info=True)
         is_success = False
 
     return is_success
@@ -70,7 +88,7 @@ def get_local_status() -> Optional[str]:
         else:
             logger.error("Failed to parse result from sg-cluster: %s", output)
     except Exception:  # noqa: BLE001
-        logger.exception("Failed to get online/offline status.")
+        logger.exception("Failed to get online/offline status.", exc_info=True)
     return status
 
 
@@ -80,11 +98,11 @@ def get_primary_status(config: HAScriptConfig, ec2: EC2ResourceType) -> str:
     this is called from secondary because it requires the config
     parameter 'primary_instance_id'. On the primary get_local_status
 
-    return 'online', 'offline' or 'unknown'
+    :param config: configuration from the main program
+    :return: 'online', 'offline' or 'unknown'
     """
     primary_instance_id = config.primary_instance_id
     if primary_instance_id is None:
-        send_error_to_smc(config, "HA not working. Config issue: " + "missing primary_instance_id")
-        sys.exit(1)
+        raise HAScriptConfigError("Config issue: missing primary_instance_id")
     status = get_config_tag_value(ec2, "status", primary_instance_id)
     return status if status is not None else "unknown"