+++ src/aws_ha_script/script.py
@@ -4,8 +4,6 @@ import logging
 import sys
 from datetime import datetime
 
-from botocore.exceptions import ClientError as BotoClientError
-
 from aws_ha_script.aws_utils import configure_ca_cert, get_config_tags, get_ec2_resource
 from aws_ha_script.config import load_config
 from aws_ha_script.daemon import (
@@ -19,9 +17,10 @@ from aws_ha_script.daemon import (
 from aws_ha_script.exceptions import HAScriptConfigError
 from aws_ha_script.log_utils import configure_logging, logger
 from aws_ha_script.mainloop import mainloop
-from aws_ha_script.smc_events import send_error_to_smc
+from aws_ha_script.ngfw_utils import is_primary
+from aws_ha_script.smc_events import send_error_to_smc, send_notification_to_smc
 
-__VERSION__ = "1.1.3"
+__VERSION__ = "1.1.4-rc2"
 
 
 def parse_args() -> argparse.Namespace:
@@ -52,40 +51,49 @@ def main() -> None:
     date_time = now.strftime("%Y%m%d")
     log_file = f"/data/diagnostics/aws-ha-{date_time}.log"
 
+    script_info = f"{ACTUAL_SCRIPT_NAME}, file: {__file__}, version: {__VERSION__}"
+
     args = parse_args()
     if args.version:
-        print(f"this is {ACTUAL_SCRIPT_NAME} version {__VERSION__}")
+        print(script_info)
         sys.exit(1)
 
     configure_logging(log_file, args.console, args.debug)
-    logger.info(
-        "%s started as %s (version %s)", ACTUAL_SCRIPT_NAME, __file__, __VERSION__
-    )
+    logger.info(f"Script started: {script_info}")
 
     die_with_parent()
     write_pid()
     install_signal_handlers()
     configure_ca_cert()
 
-    ec2 = get_ec2_resource()
-
-    config = None
+    ec2, config, role, tags = None, None, None, {}
 
     try:
+        ec2 = get_ec2_resource()
         tags = get_config_tags(ec2)
+    except Exception as exc:  # noqa: BLE001
+        logger.exception("HA not starting. Failed with exception.", exc_info=True)
+        send_error_to_smc(config, f"HA not starting. Failed to get aws tags: {exc}")
+        disable_service_and_exit(1)
+
+    try:
         config = load_config(tags)
+        role = "primary" if is_primary(config) else "secondary"
     except OSError as io_error:
-        send_error_to_smc(config, f"HA not working. failed to read config: {io_error}")
+        send_error_to_smc(config, f"HA not starting. Failed to read config: {io_error}")
         disable_service_and_exit(1)
-    except HAScriptConfigError as exc:
-        send_error_to_smc(config, f"HA not working. Invalid config: {exc}")
-        disable_service_and_exit(1)
-    except BotoClientError as exc:
-        send_error_to_smc(config, f"HA not working. Failed to get aws tags: {exc}")
+    except HAScriptConfigError as config_error:
+        send_error_to_smc(config, f"HA not starting. Invalid config: {config_error}")
         disable_service_and_exit(1)
     except Exception as exc:  # noqa: BLE001
-        logger.error("HA not working. Failed with exception: %s", exc)
+        logger.exception("HA not starting. Failed with exception.", exc_info=True)
+        send_error_to_smc(config, f"HA not starting. Script exited: {exc}")
         disable_service_and_exit(1)
+
+    send_notification_to_smc(config, f"Script started: {script_info}, role: {role}")
+
+    if config.dry_run:
+        logger.warning("DRY-RUN: No changes will be made to the system.")
 
     if config.disabled:
         logger.info("Script 'run-at-boot' is disabled. Exiting.")
@@ -98,12 +106,12 @@ def main() -> None:
         mainloop(config, ec2)
     except KeyboardInterrupt:
         logger.warning("Leaving script.")
-    except Exception:  # noqa: BLE001
-        logger.critical("Unknown exception. Exiting", exc_info=True)
-        send_error_to_smc(config, "HA not working. Script exited.")
+    except Exception as exc:  # noqa: BLE001
+        logger.exception("HA not working. Failed with exception.", exc_info=True)
+        send_error_to_smc(config, f"HA not working. Script exited: {exc}")
 
     cleanup_pid()
-    logger.info("%s terminated", __file__)
+    send_notification_to_smc(config, f"Script terminated: {script_info}, role: {role}")
 
 
 if __name__ == "__main__":