+++ tests/test_aws_utils.py
@@ -18,11 +18,16 @@ from aws_ha_script.aws_utils import (
 
 def test_set_config_tag_success() -> None:
     with patch("aws_ha_script.aws_utils.get_metadata_value") as get_metadata_value:
+        config = HAScriptConfig(
+            route_table_id="rtb-0869eb690cef8c3a6",
+            primary_instance_id="i-1234",
+            secondary_instance_id="i-2345"
+        )
         ec2 = MagicMock()
         instance_id = MagicMock()
         get_metadata_value.return_value = instance_id
 
-        assert set_config_tag(ec2, "status", "online")
+        assert set_config_tag(config, ec2, "status", "online")
         assert len(ec2.mock_calls) == 1
         ec2.create_tags.assert_called_once_with(
             Resources=[instance_id], Tags=[{"Value": "online", "Key": "FP_HA_status"}]
@@ -31,26 +36,37 @@ def test_set_config_tag_success() -> None:
 
 def test_set_config_tag_fails(caplog) -> None:
     with patch("aws_ha_script.aws_utils.get_metadata_value") as get_metadata_value:
+        config = HAScriptConfig(
+            route_table_id="rtb-0869eb690cef8c3a6",
+            primary_instance_id="i-1234",
+            secondary_instance_id="i-2345"
+        )
         ec2 = MagicMock()
         instance_id = MagicMock()
         get_metadata_value.return_value = instance_id
         ec2.create_tags.side_effect = BotoClientError(MagicMock(), "create_tags")
-        assert not set_config_tag(ec2, "status", "online")
+        assert not set_config_tag(config, ec2, "status", "online")
         assert len(ec2.mock_calls) == 1
         ec2.create_tags.assert_called_once_with(
             Resources=[instance_id], Tags=[{"Value": "online", "Key": "FP_HA_status"}]
         )
 
-        assert len(caplog.records) == 1
-        assert caplog.records[0].message == "set_config_tag failed."
+        assert len(caplog.records) == 2
+        assert caplog.records[0].message.startswith("Failed to set AWS EC2 tag")
 
 
 def test_get_config_tags_success(ec2conf: Ec2Conf) -> None:
     ec2 = ec2conf.ec2
     instance = ec2conf.instances[0]
 
+    config = HAScriptConfig(
+        route_table_id=ec2conf.protected_route_table.id,
+        primary_instance_id="i-1234",
+        secondary_instance_id="i-2345"
+    )
+
     assert instance.id
-    set_config_tag(ec2, "tag1", "value1", instance_id=instance.id)
+    set_config_tag(config, ec2, "tag1", "value1", instance_id=instance.id)
 
     ec2conf.ec2.create_tags(
         Resources=[instance.id],